+++ src/Presets/Backup/Spatiebackup.php
@@ -21,7 +21,7 @@ class Spatiebackup extends Preset
         static::info('Starts the spatie backup preset installation');
 
         static::info('Run composer and artisan commands');
-        shell_exec('composer require spatie/laravel-backup -q');
+        shell_exec('composer require spatie/laravel-backup -q -n');
         shell_exec('php artisan vendor:publish --provider="Spatie\Backup\BackupServiceProvider" -q');
 
         static::info('Change the configuration file');