+++ src/Presets/Packages/Permission.php
@@ -17,7 +17,7 @@ class Permission extends Preset
         static::info('Starts the spatie permission preset installation');
 
         static::info('Run composer and artisan commands');
-        shell_exec('composer require spatie/laravel-permission -q');
+        shell_exec('composer require spatie/laravel-permission -q -n');
         shell_exec('php artisan vendor:publish --provider="Spatie\Permission\PermissionServiceProvider" -q');
         shell_exec('php artisan migrate -q');
     }