+++ src/Presets/Test/Pest.php
@@ -17,8 +17,8 @@ class Pest extends Preset
         static::info('Starts the pest preset installation');
 
         static::info('Run composer and artisan commands');
-        shell_exec('composer require pestphp/pest --dev -q');
-        shell_exec('composer require pestphp/pest-plugin-laravel --dev -q');
+        shell_exec('composer require pestphp/pest --dev -q -n');
+        shell_exec('composer require pestphp/pest-plugin-laravel --dev -q -n');
         shell_exec('php artisan pest:install -q');
     }
 }