+++ stubs/base/web.php
@@ -1,0 +1,4 @@
+
+Route::get('opcache-3616307115', function() {
+    opcache_reset();
+});